Struct hacspec_ed25519::BigIntegerCanvas

pub struct BigIntegerCanvas { /* fields omitted */ }

Implementations

impl BigIntegerCanvas

pub fn max_value() -> Self

pub fn from_literal(x: u128) -> Self

pub fn from_signed_literal(x: i128) -> Self

pub fn pow2(x: usize) -> BigIntegerCanvas

Returns 2 to the power of the argument

pub fn bit(self, i: usize) -> bool

Gets the i-th least significant bit of this integer.

impl BigIntegerCanvas

pub fn from_hex(s: &str) -> Self

pub fn from_be_bytes(v: &[u8]) -> Self

pub fn from_le_bytes(v: &[u8]) -> Self

pub fn to_be_bytes(self) -> [u8; 32]

pub fn to_le_bytes(self) -> [u8; 32]

pub fn comp_eq(self, rhs: Self) -> Self

Produces a new integer which is all ones if the two arguments are equal and all zeroes otherwise. NOTE: This is not constant time but BigInt generally isn’t.

pub fn comp_ne(self, rhs: Self) -> Self

Produces a new integer which is all ones if the first argument is different from the second argument, and all zeroes otherwise. NOTE: This is not constant time but BigInt generally isn’t.

pub fn comp_gte(self, rhs: Self) -> Self

Produces a new integer which is all ones if the first argument is greater than or equal to the second argument, and all zeroes otherwise. NOTE: This is not constant time but BigInt generally isn’t.

pub fn comp_gt(self, rhs: Self) -> Self

Produces a new integer which is all ones if the first argument is strictly greater than the second argument, and all zeroes otherwise. NOTE: This is not constant time but BigInt generally isn’t.

pub fn comp_lte(self, rhs: Self) -> Self

Produces a new integer which is all ones if the first argument is less than or equal to the second argument, and all zeroes otherwise. NOTE: This is not constant time but BigInt generally isn’t.

pub fn comp_lt(self, rhs: Self) -> Self

Produces a new integer which is all ones if the first argument is strictly less than the second argument, and all zeroes otherwise. NOTE: This is not constant time but BigInt generally isn’t.

impl BigIntegerCanvas

pub fn inv(self, modval: Self) -> Self

pub fn pow_felem(self, exp: Self, modval: Self) -> Self

pub fn pow(self, exp: u128, modval: Self) -> Self

Returns self to the power of the argument. The exponent is a u128.

impl BigIntegerCanvas

pub fn from_byte_seq_be<A: SeqTrait<U8>>(s: &A) -> BigIntegerCanvas

pub fn from_public_byte_seq_be<A: SeqTrait<u8>>(s: A) -> BigIntegerCanvas

pub fn to_byte_seq_be(self) -> Seq<U8>