Crate hacspec_ed25519

This crate gives a specification for the EdDSA signature scheme ed25519. As there is no proper consensus for the verification algorithm, it has been implemented in multiple ways.

zcash_verify implements the ZIP-0215 standard (https://zips.z.cash/zip-0215)

ietf_cofactored_verify and ietf_cofactorless_verify both implement the ietf standard. (https://datatracker.ietf.org/doc/rfc8032). However as the standard writes:

Check the group equation [8][S]B = [8]R + [8][k]A’. It’s sufficient, but not, required to instead check [S]B = R + [k]A’.

they differ in which of the verification equations they use.

alg2_verify implements algorithm 2 from the paper https://eprint.iacr.org/2020/1244.pdf.

The different implementations and specifications differ in three ways:

Cofactored/cofactorless verification – (Using [8][S]B = [8]R + [8][H(R, A, msg)]A or [S]B = R + [H(R, A, msg)]A as the verification check).

Accepting/rejecting non-canonical encodings of points.

Accepting/rejecting small-order points.

For each of the single signature verification algorithms, their corresponding batch verification algorithm has also been implemented. However one should not use cofactorless batch verification as on some inputs it is flaky (ie. it can both accept and reject the signature non-deterministically). The reason this is included is for completeness sake, to show that it is indeed flaky, and because multiple libraries implement this batch verification (eg. the dalek library). Cofactored batch verification does not have this problem.

The paper https://eprint.iacr.org/2020/1244.pdf shows why cofactorless batch verification does not work. It also provides test vectors for edge cases and compares various implementations on these.

Examples

use hacspec_lib::*;
use hacspec_ed25519::*;

let sk = SecretKey::from_hex("9d61b19deffd5a60ba844af492ec2cc44449c5697b326919703bac031cae7f60");
// or alternatively SecretKey::from_public_slice(...);
let msg = ByteSeq::from_public_slice(b"hello world");
let sig = sign(sk, &msg);
let pk = secret_to_public(sk);
assert!(zcash_verify(pk, sig, &msg).is_ok());

Structs

BatchEntry

BigInteger

BigIntegerCanvas

BigScalar

BigScalarCanvas

CompressedEdPoint

Fixed length byte array.

Ed25519FieldElement

FieldCanvas

Scalar

ScalarCanvas

SerializedScalar

Fixed length byte array.

Signature

Fixed length byte array.

Enums

Error

Functions

alg2_verify

Algorithm 2 from https://eprint.iacr.org/2020/1244.pdf. Cofactored verification. Rejects non-canonical encoding of points. Rejects small-order points for the public key.

alg3_batch_verify

Batch verification. Algorithm 3 from https://eprint.iacr.org/2020/1244.pdf. Cofactored verification. Rejects non-canonical encoding of points. Rejects small-order points for the public key.

ietf_cofactored_batch_verify

Batch verification. Cofactored verification. Rejects non-canonical encoding of points. Allows small-order points.

ietf_cofactored_verify

Cofactored verification. Rejects non-canonical encoding of points. Allows small-order points.

ietf_cofactorless_batch_verify

Batch verification. Cofactorless verification. Rejects non-canonical encoding of points. Allows small-order points. One should not use this as it can be flaky.

ietf_cofactorless_verify

Cofactorless verification. Rejects non-canonical encoding of points. Allows small-order points.

secret_to_public

sign

Sign a message under a secret key.

zcash_batch_verify

Batch verification. Cofactored verification. Allows non-canonical encoding of points. Allows small-order points.

zcash_verify

Cofactored verification. Allows non-canonical encoding of points. Allows small-order points.

Type Definitions

PublicKey

SecretKey

VerifyResult